+++ meal-ui/src/forms/Recipe/components/HeroImage.jsx
