+++ meal-ui/src/forms/Recipe/components/index.jsx
@@ -1,1 +1,3 @@
 export { default as ViewRecipeButton } from './ViewRecipeButton';
+export { default as HeroImage } from './HeroImage';
+export { default as FormActions } from './FormActions';
